+++ Fun.Blazor.Tests/ServerTests.fs
@@ -1,5 +1,6 @@
 ﻿module Fun.Blazor.Tests.ServerTests
 
+open System
 open System.Net.Http
 open System.Threading
 open System.Threading.Tasks
@@ -41,11 +42,11 @@ let ``MapGet should work with DI`` () = task {
         makeTestServer (fun route ->
             route.MapGet(
                 "/test",
-                fun (config: IConfiguration) -> task {
+                Func<_, _>(fun (config: IConfiguration) -> task {
                     do! Task.Delay 100
                     return div { config.GetValue<string>("test") }
-                }
-            )
+                })
+            ).AddFunBlazor()
             |> ignore
         )
 
@@ -60,16 +61,12 @@ let ``MapGet should work with DI`` () = task {
 let ``MapXxx should work`` () = task {
     use server =
         makeTestServer (fun route ->
-            route.MapGet("/Get1", (fun () -> testNode)) |> ignore
-            route.MapGet("/Get2", testNode) |> ignore
-            route.MapPost("/Post1", (fun () -> testNode)) |> ignore
-            route.MapPost("/Post2", testNode) |> ignore
-            route.MapPut("/Put1", (fun () -> testNode)) |> ignore
-            route.MapPut("/Put2", testNode) |> ignore
-            route.MapPatch("/Patch1", (fun () -> testNode)) |> ignore
-            route.MapPatch("/Patch2", testNode) |> ignore
-            route.MapDelete("/Delete1", (fun () -> testNode)) |> ignore
-            route.MapDelete("/Delete2", testNode) |> ignore
+            let route = route.MapGroup("").AddFunBlazor()
+            route.MapGet("/Get1", Func<_>(fun () -> testNode)) |> ignore
+            route.MapPost("/Post1",  Func<_>(fun () -> testNode)) |> ignore
+            route.MapPut("/Put1",  Func<_>(fun () -> testNode)) |> ignore
+            route.MapPatch("/Patch1",  Func<_>(fun () -> testNode)) |> ignore
+            route.MapDelete("/Delete1",  Func<_>(fun () -> testNode)) |> ignore
         )
 
     use http = server.CreateClient()
@@ -78,15 +75,8 @@ let ``MapXxx should work`` () = task {
     let! actual = http.GetStringAsync("/Get1")
     Assert.Equal(expectedResponse, actual)
 
-    let! actual = http.GetStringAsync("/Get2")
-    Assert.Equal(expectedResponse, actual)
-
 
     let! resp = http.PostAsync("/Post1", new StringContent(""))
-    let! actual = resp.Content.ReadAsStringAsync()
-    Assert.Equal(expectedResponse, actual)
-
-    let! resp = http.PostAsync("/Post2", new StringContent(""))
     let! actual = resp.Content.ReadAsStringAsync()
     Assert.Equal(expectedResponse, actual)
 
@@ -95,25 +85,13 @@ let ``MapXxx should work`` () = task {
     let! actual = resp.Content.ReadAsStringAsync()
     Assert.Equal(expectedResponse, actual)
 
-    let! resp = http.PutAsync("/Put2", new StringContent(""))
-    let! actual = resp.Content.ReadAsStringAsync()
-    Assert.Equal(expectedResponse, actual)
-
 
     let! resp = http.PatchAsync("/Patch1", new StringContent(""))
     let! actual = resp.Content.ReadAsStringAsync()
     Assert.Equal(expectedResponse, actual)
 
-    let! resp = http.PatchAsync("/Patch2", new StringContent(""))
-    let! actual = resp.Content.ReadAsStringAsync()
-    Assert.Equal(expectedResponse, actual)
-
 
     let! resp = http.DeleteAsync("/Delete1")
-    let! actual = resp.Content.ReadAsStringAsync()
-    Assert.Equal(expectedResponse, actual)
-
-    let! resp = http.DeleteAsync("/Delete2")
     let! actual = resp.Content.ReadAsStringAsync()
     Assert.Equal(expectedResponse, actual)
 }